+++ tidy3d/plugins/invdes/__init__.py
@@ -4,6 +4,7 @@ from . import utils
 from .design import InverseDesign, InverseDesignMulti
 from .optimizer import AdamOptimizer
 from .penalty import ErosionDilationPenalty
+from .postprocess import GetPowerMode, WeightedSum
 from .region import TopologyDesignRegion
 from .result import InverseDesignResult
 from .transformation import FilterProject
@@ -16,5 +17,7 @@ __all__ = (
     "TopologyDesignRegion",
     "AdamOptimizer",
     "InverseDesignResult",
+    "WeightedSum",
+    "GetPowerMode",
     "utils",
 )